# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hoddmímir API",
    "version": "1.0.0",
    "description": "Authenticated, role-protected API for read projections and verified Hoddmímir configuration mutations."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "sessionCookie": [] }],
  "paths": {
    "/api/v1/operations/pbs-tasks": {
      "get": {
        "operationId": "listPbsTasks",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9999999,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Persisted collector observations; inspection freshness and truncation are explicit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PbsObservedTaskPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidQuery"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "503": {
            "$ref": "#/components/responses/ReadModelUnavailable"
          }
        }
      }
    },
    "/api/v1/operations/pbs-tasks/{id}": {
      "get": {
        "operationId": "getPbsTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Persisted collector observations; inspection freshness and truncation are explicit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PbsObservedTaskDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidQuery"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "503": {
            "$ref": "#/components/responses/ReadModelUnavailable"
          },
          "404": {
            "description": "PBS task not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object", "additionalProperties": false, "required": ["error"], "properties": {"error": {"type": "object", "additionalProperties": false, "required": ["code"], "properties": {"code": {"const": "pbs_task_not_found"}}}}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/shadow/evaluations": {
      "get": {
        "operationId": "listShadowEvaluations",
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Cursor" }
        ],
        "responses": {
          "200": {
            "description": "Persisted read-only shadow evaluation runs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowEvaluationPage"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "503": { "$ref": "#/components/responses/ReadModelUnavailable" }
        }
      }
    },
    "/api/v1/shadow/decisions": {
      "get": {
        "operationId": "listShadowDecisions",
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Cursor" },
          {
            "name": "outcome",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ShadowOutcome" }
          },
          {
            "name": "reason",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ShadowReason" }
          },
          {
            "name": "policyId",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
          },
          {
            "name": "targetId",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
          },
          {
            "name": "guestId",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Persisted read-only shadow decisions.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShadowDecisionPage" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "503": { "$ref": "#/components/responses/ReadModelUnavailable" }
        }
      }
    },
    "/api/v1/shadow/decisions/{id}": {
      "get": {
        "operationId": "getShadowDecision",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "One shadow decision with its ordered gates.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowDecisionDetail"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "404": { "description": "The shadow decision was not found." },
          "503": { "$ref": "#/components/responses/ReadModelUnavailable" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hoddmimir_session"
      }
    },
    "parameters": {
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/PageCursor" }
      },
...
    },
    "responses": {
...
      "AuthenticationRequired": {
        "description": "A valid local web session is required.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AuthError" }
          }
        }
      },
      "PermissionDenied": {
        "description": "Permission or CSRF validation failed.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AuthError" }
          }
        }
      },
      "InvalidQuery": {
        "description": "Unknown, malformed, incompatible, or out-of-bounds query.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
      "ReadModelUnavailable": {
        "description": "The read-only projection is temporarily unavailable without exposing database or runtime details.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [{ "$ref": "#/components/schemas/ReadModelUnavailableError" }, { "$ref": "#/components/schemas/MaintenanceError" }]
            }
          }
        }
      },
...
    },
    "schemas": {
      "PbsObservedTask": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "connectionId",
          "workerType",
          "workerId",
          "lifecycle",
          "remoteStatus",
          "startedAt",
          "lastSeenAt",
          "inspectedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "connectionId": {
            "type": "string",
            "format": "uuid"
          },
          "workerType": {
            "type": "string"
          },
          "workerId": {
            "type": [
              "string",
              "null"
            ]
          },
          "lifecycle": {
            "type": "string",
            "enum": [
              "running",
              "stopped",
              "unknown"
            ]
          },
          "remoteStatus": {
            "type": [
              "string",
              "null"
            ]
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "inspectedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "PbsObservedTaskPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "$ref": "#/components/schemas/PbsObservedTask"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "PbsObservedTaskDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "connectionId",
          "workerType",
          "workerId",
          "lifecycle",
          "remoteStatus",
          "startedAt",
          "lastSeenAt",
          "inspectedAt",
          "upid",
          "inspection"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "connectionId": {
            "type": "string",
            "format": "uuid"
          },
          "workerType": {
            "type": "string"
          },
          "workerId": {
            "type": [
              "string",
              "null"
            ]
          },
          "lifecycle": {
            "type": "string",
            "enum": [
              "running",
              "stopped",
              "unknown"
            ]
          },
          "remoteStatus": {
            "type": [
              "string",
              "null"
            ]
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "inspectedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "upid": {
            "type": "string"
          },
          "inspection": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": false,
            "required": [
              "status",
              "exitStatus",
              "endTime",
              "lines",
              "truncated",
              "statusFailure",
              "logFailure"
            ],
            "properties": {
              "status": {
                "type": [
                  "string",
                  "null"
                ],
                "enum": [
                  "running",
                  "stopped",
                  null
                ]
              },
              "exitStatus": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "endTime": {
                "type": [
                  "integer",
                  "null"
                ]
              },
              "lines": {
                "type": "array",
                "maxItems": 500,
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "number",
                    "text"
                  ],
                  "properties": {
                    "number": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "text": {
                      "type": "string"
                    }
                  }
                }
              },
              "truncated": {
                "type": "boolean"
              },
              "statusFailure": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "logFailure": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          }
        }
      },
...
      "MaintenanceError": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": {
          "type": "object", "additionalProperties": false, "required": ["code", "message"],
          "properties": {
            "code": { "const": "maintenance" },
            "message": { "const": "Hoddmímir wird gewartet." }
          }
        } }
      },
...
      "AuthError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "authentication_required",
                  "authentication_failed",
                  "permission_denied"
                ]
              }
            }
          }
        }
      },
...
      "CanonicalUuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      },
      "UtcTimestamp": {
        "type": "string",
        "format": "date-time",
        "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\\.[0-9]{6}Z$"
      },
...
      "PageCursor": {
        "type": "string",
        "minLength": 1,
        "maxLength": 2048,
        "pattern": "^[A-Za-z0-9_-]+$"
      },
...
      "PageMetadata": {
        "type": "object",
        "additionalProperties": false,
        "required": ["limit", "count", "hasMore", "nextCursor"],
        "properties": {
          "limit": { "type": "integer", "minimum": 1, "maximum": 100 },
          "count": { "type": "integer", "minimum": 0, "maximum": 100 },
          "hasMore": { "type": "boolean" },
          "nextCursor": {
            "oneOf": [
              { "$ref": "#/components/schemas/PageCursor" },
              { "type": "null" }
            ]
          }
        }
      },
...
      "ShadowOutcome": {
        "type": "string",
        "enum": ["eligible", "blocked", "not_due", "deduplicated"]
      },
      "ShadowReason": {
        "type": "string",
        "enum": ["manual", "never_backed_up", "max_age", "bytes_written"]
      },
      "ShadowGateCode": {
        "type": "string",
        "enum": [
          "connection_enabled",
          "cluster_enabled",
          "node_enabled",
          "guest_enabled",
          "policy_enabled",
          "policy_failure_notification_configured",
          "policy_retention_compatible",
          "target_enabled",
          "explicit_exclusion_absent",
          "guest_active",
          "inventory_fresh",
          "placement_present",
          "placement_fresh",
          "active_request_absent",
          "target_node_allowed",
          "target_storage_enabled",
          "target_storage_active",
          "executor_authorization_fresh",
          "executor_authorized",
          "capacity_fresh",
          "minimum_free_space",
          "node_concurrency",
          "target_concurrency",
          "pbs_mapping_valid",
          "higher_ranked_candidate_absent"
        ]
      },
      "ShadowGateScope": {
        "type": "string",
        "enum": [
          "connection",
          "cluster",
          "node",
          "guest",
          "policy",
          "target",
          "inventory",
          "placement",
          "authorization",
          "capacity",
          "concurrency",
          "request",
          "pbs_mapping"
        ]
      },
      "ShadowGateDetailCode": {
        "type": "string",
        "enum": [
          "passed",
          "disabled",
          "explicitly_excluded",
          "archived",
          "missing",
          "stale",
          "not_allowed",
          "inactive",
          "unauthorized",
          "insufficient_free_space",
          "concurrency_limit_reached",
          "invalid_mapping",
          "incompatible",
          "unconfigured",
          "active_request_exists",
          "higher_ranked_candidate"
        ]
      },
      "ShadowEvaluation": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "cycleToken",
          "fencingToken",
          "evaluatorVersion",
          "decisionCount",
          "gateCount",
          "startedAt",
          "completedAt",
          "persistedAt"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/CanonicalUuid" },
          "cycleToken": { "$ref": "#/components/schemas/CanonicalUuid" },
          "fencingToken": { "type": "integer", "minimum": 1 },
          "evaluatorVersion": { "type": "integer", "minimum": 1 },
          "decisionCount": { "type": "integer", "minimum": 0 },
          "gateCount": { "type": "integer", "minimum": 0 },
          "startedAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "completedAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "persistedAt": { "$ref": "#/components/schemas/UtcTimestamp" }
        }
      },
      "ShadowDecision": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "evaluationId",
          "guestId",
          "nodeId",
          "outcome",
          "reason",
          "priority",
          "policyId",
          "policyRevision",
          "targetId",
          "targetRevision",
          "completedAt"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/CanonicalUuid" },
          "evaluationId": { "$ref": "#/components/schemas/CanonicalUuid" },
          "guestId": { "$ref": "#/components/schemas/CanonicalUuid" },
          "nodeId": {
            "oneOf": [
              { "$ref": "#/components/schemas/CanonicalUuid" },
              { "type": "null" }
            ]
          },
          "outcome": { "$ref": "#/components/schemas/ShadowOutcome" },
          "reason": {
            "oneOf": [
              { "$ref": "#/components/schemas/ShadowReason" },
              { "type": "null" }
            ]
          },
          "priority": {
            "oneOf": [
              { "type": "integer", "enum": [100, 200, 300, 400] },
              { "type": "null" }
            ]
          },
          "policyId": { "$ref": "#/components/schemas/CanonicalUuid" },
          "policyRevision": { "type": "integer", "minimum": 1 },
          "targetId": { "$ref": "#/components/schemas/CanonicalUuid" },
          "targetRevision": { "type": "integer", "minimum": 1 },
          "completedAt": { "$ref": "#/components/schemas/UtcTimestamp" }
        }
      },
      "ShadowGate": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "position",
          "code",
          "passed",
          "scope",
          "subjectId",
          "observedAt",
          "detailCode"
        ],
        "properties": {
          "position": { "type": "integer", "minimum": 1 },
          "code": { "$ref": "#/components/schemas/ShadowGateCode" },
          "passed": { "type": "boolean" },
          "scope": { "$ref": "#/components/schemas/ShadowGateScope" },
          "subjectId": { "$ref": "#/components/schemas/CanonicalUuid" },
          "observedAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          },
          "detailCode": { "$ref": "#/components/schemas/ShadowGateDetailCode" }
        }
      },
      "ShadowDecisionDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/ShadowDecision" },
          {
            "type": "object",
            "required": ["gates"],
            "properties": {
              "gates": {
                "type": "array",
                "minItems": 1,
                "items": { "$ref": "#/components/schemas/ShadowGate" }
              }
            }
          }
        ],
        "unevaluatedProperties": false
      },
      "ShadowEvaluationPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page"],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": { "$ref": "#/components/schemas/ShadowEvaluation" }
          },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        }
      },
      "ShadowDecisionPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page"],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": { "$ref": "#/components/schemas/ShadowDecision" }
          },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        }
      },
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message"],
            "properties": {
              "code": { "const": "invalid_query" },
              "message": { "const": "The query is invalid." }
            }
          }
        }
      },
      "ReadModelUnavailableError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message"],
            "properties": {
              "code": { "const": "read_model_unavailable" },
              "message": {
                "const": "The read model is temporarily unavailable."
              }
            }
          }
        }
      }
    }
  }
}